fix(enhance_advanced): 스켈레톤 빌드 차단 이슈 6a-6g 해소 및 BUILD_ON 복원
2026-04-18 감사에서 차기 스프린트로 이월된 6a-6d 를 모두 해소하고
신규 발견된 6e-6g (LNK2005 version 중복, GTest FetchContent 누락) 까지
포함하여 xpe_enhance_advanced.dll 의 독립 빌드를 복원한다.

## 해소 내역
- 6a: xpe_enhance_advanced.cpp 의 상대 include 경로 정정 ("fractional_derivative.h"
  / "exposure_index.h" → "detail/fractional_derivative.h" / "detail/exposure_index.h").
- 6b: XpeErrorCode 사용 헤더 (mfp_scalar.h, fractional_derivative.h) 에
  xpe_error.h include 보강.
- 6c: Eigen::Index → int, M_PI/M_PI_2 (double) → float narrowing 을 모든
  호출 지점에서 명시적 static_cast 로 정리. /WX 체제에서 모든 C4244
  경고가 오류로 승격되지 않도록 함.
- 6d: 내부 헤더(fractional_derivative.h, exposure_index.h) 를
  include/xpe/enhance_advanced/ 공개 경로에서 src/detail/ 내부 경로로
  이동하여 ABI 경계 위생 확보. 네임스페이스(xpe::enhance_advanced) 는 유지.
- 6e: CMake `create_placeholder_version_function` 이 실제 구현과 중복
  정의(LNK2005)를 유발 → 제거. 실 구현은 xpe_enhance_advanced.cpp 에
  존재.
- 6f (부분): mfp_scalar.cpp size_t → int, xpe_enhance_advanced.cpp
  unused catch variable (C4101/C4267) 제거.
- 6g: enhance_advanced/CMakeLists.txt 가 GTest 를 find_package 만 시도해
  monorepo 빌드에서 실패 → modules/common 패턴 미러링하여
  FetchContent 폴백 추가.

## 빌드 옵션 복원
- 루트 CMakeLists.txt: BUILD_ENHANCE_ADVANCED 기본값 OFF → ON.
- 모듈 CMakeLists.txt: target_compile_options /W4 /WX (MSVC) 및
  -Wall -Wextra -Werror (비-MSVC) per-target 적용하여 모듈 수준
  warnings-as-errors 체제 가동.

## 검증
- cmake -B build/ci-enhance-advanced -G "Visual Studio 17 2022" -A x64 ...
  configure 성공.
- msbuild Release 빌드 성공: xpe_enhance_advanced.dll (142 KB) 생성.
- /WX 체제에서 모듈 대상 경고 0 달성.

SPEC: SPEC-XPE-P2-ADV
🗿 MoAI <email@mo.ai.kr>

+++ modules/enhance_advanced/src/mfp_scalar.cpp
@@ -123,7 +123,7 @@ void LaplacianPyramid::reconstruct(const MfpConfig& config, float* outData) {
         std::vector<float> upsampled(w * h);
         int currentW = reconstructed.size() > 0 ?
             static_cast<int>(std::sqrt(reconstructed.size())) : 1;
-        int currentH = reconstructed.size() / currentW;
+        int currentH = static_cast<int>(reconstructed.size() / static_cast<size_t>(currentW));
 
         upsample(reconstructed.data(), upsampled.data(), currentW, currentH);
